+++ rs-application/application-api/src/routes.rs
@@ -1,9 +1,8 @@
 use crate::middleware::authorization;
 use crate::v1;
-use salvo::prelude::StatusCode;
-use salvo::{Depot, FlowCtrl, Request, Response, Router, handler};
+use salvo::prelude::{Json, StatusCode};
+use salvo::{handler, Depot, FlowCtrl, Request, Response, Router};
 
-#[handler]
 #[handler]
 pub fn catcher(_req: &Request, _depot: &Depot, res: &mut Response, ctrl: &mut FlowCtrl) {
     let (code, msg) = match res.status_code {
@@ -12,11 +11,11 @@ pub fn catcher(_req: &Request, _depot: &Depot, res: &mut Response, ctrl: &mut Fl
         _ => return,
     };
 
-    res.render(crate::response::Response::<String>::new(
+    res.render(Json(crate::response::Response::<String>::new(
         Some(code),
         Some(msg.to_string()),
         None,
-    ));
+    )));
 
     ctrl.skip_rest();
 }